Move translate's JSONs to root
