+++ src/operators.json
